refactor: unify SqlMessageCatalogue into InMemoryMessageCatalogue
SqlMessageCatalogue was functionally identical to InMemoryMessageCatalogue -
both just wrap a Map[String, String] for message lookup. The "Sql" prefix
was misleading since the class itself had no SQL-specific code.

Changes:
- SqlMessageCatalogueService now creates InMemoryMessageCatalogue instances
- Enhanced InMemoryMessageCatalogue with comprehensive ScalaDoc
- Removed redundant assume() checks from InMemoryMessageCatalogue
- Deleted SqlMessageCatalogue.scala
- Deleted MessageCatalogueComparisonSpec.scala (no longer needed)
- Created InMemoryMessageCatalogueSpec.scala with unit tests
- Updated references in MessageCatalogue.scala and test files

The architecture is now cleaner:
- InMemoryMessageCatalogue: data-source agnostic Map-backed catalogue
- SqlMessageCatalogueService: SQL-specific service managing DB loading
