+++ core/jvm/src/main/scala/works/iterative/core/service/impl/InMemoryMessageCatalogue.scala
@@ -1,3 +1,6 @@
+// PURPOSE: MessageCatalogue implementation backed by pre-loaded in-memory Map
+// PURPOSE: Provides pure synchronous access to messages without any IO during message lookup
+
 package works.iterative
 package core.service.impl
 
@@ -8,22 +11,40 @@ import zio.*
 import zio.json.*
 import works.iterative.core.Language
 
+/** MessageCatalogue implementation backed by pre-loaded in-memory Map.
+  *
+  * This implementation provides synchronous access to messages stored in an immutable Map.
+  * Messages are typically loaded once (at startup or from resources) and stored for fast O(1) lookups.
+  *
+  * Key characteristics:
+  *   - Pure synchronous access (no effects during message retrieval)
+  *   - O(1) lookup performance using immutable Map
+  *   - Thread-safe (immutable data structure)
+  *   - Formatting errors return error message instead of throwing exceptions
+  *
+  * Usage:
+  *   - With JSON resources: Use companion object's `fromJsonResources` method
+  *   - With SQL database: Used by [[SqlMessageCatalogueService]] which manages pre-load and reload lifecycle
+  *
+  * @param language The language for this message catalogue
+  * @param messages Pre-loaded map of message keys to message text
+  *
+  * @see [[SqlMessageCatalogueService]] for database-backed message catalogue service
+  * @see [[InMemoryMessageCatalogueService]] for JSON resource-backed service
+  */
 class InMemoryMessageCatalogue(override val language: Language, messages: Map[String, String])
     extends MessageCatalogue:
 
     override def get(id: MessageId): Option[String] =
-        assume(messages != null, "Message catalogue must not be null")
         messages.get(id.toString)
 
     override def get(msg: UserMessage): Option[String] =
-        assume(messages != null, "Message catalogue must not be null")
-        get(msg.id).map(m =>
-            Try(m.format(msg.args*)).fold(
-                t => s"error formatting [${msg.id.toString()}]: '$m': ${t.getMessage}",
+        get(msg.id).map(template =>
+            Try(template.format(msg.args*)).fold(
+                exception => s"error formatting [${msg.id}]: '$template': ${exception.getMessage}",
                 identity
             )
         )
-    end get
 
     override val root: MessageCatalogue = this
 end InMemoryMessageCatalogue
